requirements.txt
pandas==1.5.2
SQLAlchemy==1.4.44
psycopg2~=2.9.3
PyYAML~=6.0
openpyxl

requests~=2.28.1
tqdm~=4.64.1
urllib3==1.26.13
uvicorn~=0.20.0
fastapi==0.88.0
pydantic~=1.10.2
pyodbc==4.0.35
charset-normalizer
chardet
lxml